Incremental index as workflow
