+++ graphrag/index/run/utils.py
@@ -9,12 +9,14 @@ from graphrag.callbacks.noop_workflow_callbacks import NoopWorkflowCallbacks
 from graphrag.callbacks.progress_workflow_callbacks import ProgressWorkflowCallbacks
 from graphrag.callbacks.workflow_callbacks import WorkflowCallbacks
 from graphrag.callbacks.workflow_callbacks_manager import WorkflowCallbacksManager
+from graphrag.config.models.graph_rag_config import GraphRagConfig
 from graphrag.index.typing.context import PipelineRunContext
 from graphrag.index.typing.state import PipelineState
 from graphrag.index.typing.stats import PipelineRunStats
 from graphrag.logger.base import ProgressLogger
 from graphrag.storage.memory_pipeline_storage import MemoryPipelineStorage
 from graphrag.storage.pipeline_storage import PipelineStorage
+from graphrag.utils.api import create_storage_from_config
 
 
 def create_run_context(
@@ -44,3 +46,16 @@ def create_callback_chain(
     if progress is not None:
         manager.register(ProgressWorkflowCallbacks(progress))
     return manager
+
+
+def get_update_storages(
+    config: GraphRagConfig, timestamp: str
+) -> tuple[PipelineStorage, PipelineStorage, PipelineStorage]:
+    """Get storage objects for the update index run."""
+    output_storage = create_storage_from_config(config.output)
+    update_storage = create_storage_from_config(config.update_index_output)
+    timestamped_storage = update_storage.child(timestamp)
+    delta_storage = timestamped_storage.child("delta")
+    previous_storage = timestamped_storage.child("previous")
+
+    return output_storage, previous_storage, delta_storage
