+++ graphrag/index/workflows/__init__.py
@@ -42,6 +42,27 @@ from .generate_text_embeddings import (
 from .prune_graph import (
     run_workflow as run_prune_graph,
 )
+from .update_communities import (
+    run_workflow as run_update_communities,
+)
+from .update_community_reports import (
+    run_workflow as run_update_community_reports,
+)
+from .update_covariates import (
+    run_workflow as run_update_covariates,
+)
+from .update_entities_relationships import (
+    run_workflow as run_update_entities_relationships,
+)
+from .update_final_documents import (
+    run_workflow as run_update_final_documents,
+)
+from .update_text_embeddings import (
+    run_workflow as run_update_text_embeddings,
+)
+from .update_text_units import (
+    run_workflow as run_update_text_units,
+)
 
 # register all of our built-in workflows at once
 PipelineFactory.register_all({
@@ -57,4 +78,11 @@ PipelineFactory.register_all({
     "finalize_graph": run_finalize_graph,
     "generate_text_embeddings": run_generate_text_embeddings,
     "prune_graph": run_prune_graph,
+    "update_final_documents": run_update_final_documents,
+    "update_text_embeddings": run_update_text_embeddings,
+    "update_community_reports": run_update_community_reports,
+    "update_entities_relationships": run_update_entities_relationships,
+    "update_communities": run_update_communities,
+    "update_covariates": run_update_covariates,
+    "update_text_units": run_update_text_units,
 })
